refact: add folder to organize files into draft and professional
