+++ src/main/kotlin/com/effies/draft/application/services/TeamService.kt
@@ -3,16 +3,16 @@ package com.effies.draft.application.services
 import com.effies.draft.application.exceptions.AlreadyExistsException
 import com.effies.draft.application.exceptions.NotFoundException
 import com.effies.draft.application.exceptions.ValidatorException
-import com.effies.draft.mappers.toDomain
-import com.effies.draft.mappers.toEntity
+import com.effies.draft.mappers.draft.toDomain
+import com.effies.draft.mappers.draft.toEntity
 import com.effies.draft.application.port.out.repositories.FinancialRepository
 import com.effies.draft.application.port.out.repositories.ScoreRepository
 import com.effies.draft.application.port.out.repositories.TeamRepository
 import com.effies.draft.application.validations.TeamValidator
-import com.effies.draft.domains.Financial
-import com.effies.draft.domains.Score
-import com.effies.draft.domains.Team
-import com.effies.draft.domains.TeamStats
+import com.effies.draft.domains.draft.Financial
+import com.effies.draft.domains.draft.Score
+import com.effies.draft.domains.draft.Team
+import com.effies.draft.domains.draft.TeamStats
 import java.util.UUID
 
 class TeamService(
