+++ src/main/kotlin/com/effies/draft/domains/draft/Financial.kt
@@ -1,4 +1,4 @@
-package com.effies.draft.domains
+package com.effies.draft.domains.draft
 
 data class Financial(
     val budget: Double = 0.00,
