+++ src/main/kotlin/com/effies/draft/domains/professional/ProTeams.kt
@@ -1,4 +1,4 @@
-package com.effies.draft.domains
+package com.effies.draft.domains.professional
 
 data class ProTeams(
     val teams: List<ProTeam>
